refactor: [Task 1.7] Extract LoadingIndicator to dedicated module
